+++ src/libcore/fmt/rt.rs
